requirements.txt
pytest==7.4.4
pytest-reportportal==5.3.1
# to retry flaky tests. pytest --retries 2
pytest-retry==1.6.2

# UI libs
playwright==1.43.0

# API libs
requests==2.31.0
requests-to-curl==1.1.0

# OTHER
pydantic==2.6.4
jsonpickle==3.0.2
python-configuration==0.9.1
assertpy==1.1
streams.py==1.2.1